+++ jgclark.Dashboard/src/dashboardHelpers.js
@@ -24,6 +24,7 @@ import {
   stripTodaysDateRefsFromString
 } from '@helpers/stringTransforms'
 import { showMessage } from '@helpers/userInput'
+import { showMessageYesNo } from '../../helpers/userInput'
 
 //-----------------------------------------------------------------
 // Data types
@@ -54,8 +55,10 @@ export type dashboardConfigType = {
   ignoreTasksWithPhrase: string,
   includeFolderName: boolean,
   includeTaskContext: boolean,
+  windowWidth: number,
+  windowHeight: number,
   _logLevel: string,
-  _triggerLogging: boolean,
+  triggerLogging: boolean,
 }
 
 /**
@@ -88,12 +91,14 @@ export async function getSettings(): Promise<any> {
 export async function checkForRequiredSharedFiles(): Promise<void> {
   // logDebug(pluginJson, `Start of checkForRequiredSharedFiles()`)
   try {
+    const sharedPluginID = "np.Shared"
+    const sharedPluginName = "Shared Resources"
     const wantedFileList = pluginJson['plugin.requiredSharedFiles']
-    logDebug(`${pluginID}/init`, `${String(wantedFileList.length)} wanted files: ${String(wantedFileList)}`)
+    logDebug(`dashboard/helpers/checkForRequiredSharedFiles`, `${String(wantedFileList.length)} wanted files: ${String(wantedFileList)}`)
     const wantedRes = await checkForWantedResources(wantedFileList)
     if (typeof wantedRes === 'number' && wantedRes >= wantedFileList.length) {
       // plugin np.Shared is loaded, and is providing all the wanted resources
-      logDebug(`${pluginID}/init`, `plugin np.Shared is loaded 😄 and provides all the ${String(wantedFileList.length)} wanted files`)
+      logDebug(`dashboard/helpers/checkForRequiredSharedFiles`, `plugin np.Shared is loaded 😄 and provides all the ${String(wantedFileList.length)} wanted files`)
     } else if (typeof wantedRes === 'number' && wantedRes < wantedFileList.length) {
       // plugin np.Shared is loaded, but isn't providing all the wanted resources
       logWarn(
@@ -102,10 +107,21 @@ export async function checkForRequiredSharedFiles(): Promise<void> {
       )
     } else if (wantedRes) {
       // plugin np.Shared is loaded
-      logDebug(`${pluginID}/init`, `plugin np.Shared is loaded 😄; no further checking done`)
+      logDebug(`dashboard/helpers/checkForRequiredSharedFiles`, `plugin np.Shared is loaded 😄; no further checking done`)
     } else {
       // plugin np.Shared is not loaded
-      logWarn(`${pluginID}/init`, `plugin np.Shared isn't loaded 🥵, so icons probably won't display`)
+      logWarn(`dashboard/helpers/checkForRequiredSharedFiles`, `plugin np.Shared isn't loaded 🥵, so icons probably won't display`)
+      const res = await showMessageYesNo(`It looks like you haven't installed the '${sharedPluginName}' plugin, which is required for the Dashboard to operate properly. Would you like me to install it for you?`, ['Yes', 'No'], 'Dashboard plugin problem')
+      if (res === 'Yes') {
+        const pluginObjects = await DataStore.listPlugins(true) ?? []
+        const pluginObject = pluginObjects?.find((p) => p.id === sharedPluginID)
+        if (pluginObject) {
+          // clo(pluginObject, `installPlugin "${sharedPluginID}"`)
+          await DataStore.installPlugin(pluginObject, true)
+        } else {
+          await showMessage(`Could not install '${sharedPluginName}' plugin: not found in repository. If this persists, please contact NotePlan support.`)
+        }
+      }
     }
   } catch (err) {
     logError(pluginJson, `${err.name}: ${err.message}`)
@@ -253,6 +269,7 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
     // (Simpler regex possible as the count comes later)
     // Note: this wrapping needs to go last
     if (output.match(/\B\!+\B/)) {
+      // $FlowFixMe(incompatible-use)
       const numExclamations = output.match(/\B\!+\B/)[0].length
       switch (numExclamations) {
         case 1: {